Move publish api content
