+++ cmd/exposure/main.go
@@ -21,8 +21,8 @@ import (
 	"net/http"
 
 	"github.com/google/exposure-notifications-server/internal/api/handlers"
-	"github.com/google/exposure-notifications-server/internal/api/publish"
 	"github.com/google/exposure-notifications-server/internal/logging"
+	"github.com/google/exposure-notifications-server/internal/publish"
 	"github.com/google/exposure-notifications-server/internal/setup"
 )
